Update progress to match 5x branch
