+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/Exceptions/PopInvalidAtClaimException.cs
@@ -27,7 +27,7 @@
 
 using System;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
     /// </summary>
