+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/Exceptions/PopInvalidCnfClaimException.cs
@@ -27,34 +27,34 @@
 
 using System;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
     /// </summary>
-    public class PopInvalidMClaimException : PopValidationException
+    public class PopInvalidCnfClaimException : PopValidationException
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidMClaimException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidCnfClaimException"/> class.
         /// </summary>
-        public PopInvalidMClaimException()
+        public PopInvalidCnfClaimException()
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidMClaimException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidCnfClaimException"/> class.
         /// </summary>
         /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
-        public PopInvalidMClaimException(string message)
+        public PopInvalidCnfClaimException(string message)
             : base(message)
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidMClaimException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidCnfClaimException"/> class.
         /// </summary>
         /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
         /// <param name="innerException">A <see cref="Exception"/> that represents the root cause of the exception.</param>
-        public PopInvalidMClaimException(string message, Exception innerException)
+        public PopInvalidCnfClaimException(string message, Exception innerException)
             : base(message, innerException)
         {
         }
