+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/Exceptions/PopInvalidMClaimException.cs
@@ -27,34 +27,34 @@
 
 using System;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
     /// </summary>
-    public class PopInvalidPopKeyException : PopValidationException
+    public class PopInvalidMClaimException : PopValidationException
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidPopKeyException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidMClaimException"/> class.
         /// </summary>
-        public PopInvalidPopKeyException()
+        public PopInvalidMClaimException()
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidPopKeyException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidMClaimException"/> class.
         /// </summary>
         /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
-        public PopInvalidPopKeyException(string message)
+        public PopInvalidMClaimException(string message)
             : base(message)
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidPopKeyException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidMClaimException"/> class.
         /// </summary>
         /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
         /// <param name="innerException">A <see cref="Exception"/> that represents the root cause of the exception.</param>
-        public PopInvalidPopKeyException(string message, Exception innerException)
+        public PopInvalidMClaimException(string message, Exception innerException)
             : base(message, innerException)
         {
         }
