+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/Exceptions/PopInvalidPClaimException.cs
@@ -25,33 +25,38 @@
 //
 //------------------------------------------------------------------------------
 
-using Microsoft.IdentityModel.Tokens;
+using System;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
-    /// 
     /// </summary>
-    public class PopTokenValidationResult
+    public class PopInvalidPClaimException : PopValidationException
     {
         /// <summary>
-        /// Gets or sets the AccessToken.
+        /// Initializes a new instance of the <see cref="PopInvalidPClaimException"/> class.
         /// </summary>
-        public string AccessToken { get; set; }
+        public PopInvalidPClaimException()
+        {
+        }
 
         /// <summary>
-        /// Gets or sets the PopToken.
+        /// Initializes a new instance of the <see cref="PopInvalidPClaimException"/> class.
         /// </summary>
-        public string PopToken { get; set; }
+        /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
+        public PopInvalidPClaimException(string message)
+            : base(message)
+        {
+        }
 
         /// <summary>
-        /// Gets or sets the validated access token.
+        /// Initializes a new instance of the <see cref="PopInvalidPClaimException"/> class.
         /// </summary>
-        public SecurityToken ValidatedAccessToken { get; set; }
-
-        /// <summary>
-        /// Gets or sets the validated PopToken.
-        /// </summary>
-        public SecurityToken ValidatedPopToken { get; set; }
+        /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
+        /// <param name="innerException">A <see cref="Exception"/> that represents the root cause of the exception.</param>
+        public PopInvalidPClaimException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
     }
 }
