+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/Exceptions/PopInvalidPopKeyException.cs
@@ -27,34 +27,34 @@
 
 using System;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
     /// </summary>
-    public class PopInvalidPClaimException : PopValidationException
+    public class PopInvalidPopKeyException : PopValidationException
     {
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidPClaimException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidPopKeyException"/> class.
         /// </summary>
-        public PopInvalidPClaimException()
+        public PopInvalidPopKeyException()
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidPClaimException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidPopKeyException"/> class.
         /// </summary>
         /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
-        public PopInvalidPClaimException(string message)
+        public PopInvalidPopKeyException(string message)
             : base(message)
         {
         }
 
         /// <summary>
-        /// Initializes a new instance of the <see cref="PopInvalidPClaimException"/> class.
+        /// Initializes a new instance of the <see cref="PopInvalidPopKeyException"/> class.
         /// </summary>
         /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
         /// <param name="innerException">A <see cref="Exception"/> that represents the root cause of the exception.</param>
-        public PopInvalidPClaimException(string message, Exception innerException)
+        public PopInvalidPopKeyException(string message, Exception innerException)
             : base(message, innerException)
         {
         }
