+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/Exceptions/PopInvalidSignatureException.cs
@@ -23,27 +23,40 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 //
+//------------------------------------------------------------------------------
 
-using Microsoft.IdentityModel.Tokens;
-using System.Threading;
-using System.Threading.Tasks;
+using System;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
-    /// 
     /// </summary>
-    public interface IPopTokenValidator
+    public class PopInvalidSignatureException : PopValidationException
     {
         /// <summary>
-        /// 
+        /// Initializes a new instance of the <see cref="PopInvalidSignatureException"/> class.
         /// </summary>
-        /// <param name="popToken"></param>
-        /// <param name="httpRequestData"></param>
-        /// <param name="tokenValidationParameters"></param>
-        /// <param name="popTokenValidationPolicy"></param>
-        /// <param name="cancellationToken"></param>
-        /// <returns></returns>
-        Task<PopTokenValidationResult> ValidatePopTokenAsync(string popToken, HttpRequestData httpRequestData, TokenValidationParameters tokenValidationParameters, PopTokenValidationPolicy popTokenValidationPolicy, CancellationToken cancellationToken);
+        public PopInvalidSignatureException()
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="PopInvalidSignatureException"/> class.
+        /// </summary>
+        /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
+        public PopInvalidSignatureException(string message)
+            : base(message)
+        {
+        }
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="PopInvalidSignatureException"/> class.
+        /// </summary>
+        /// <param name="message">Additional information to be included in the exception and displayed to user.</param>
+        /// <param name="innerException">A <see cref="Exception"/> that represents the root cause of the exception.</param>
+        public PopInvalidSignatureException(string message, Exception innerException)
+            : base(message, innerException)
+        {
+        }
     }
 }
