+++ src/Microsoft.IdentityModel.Protocols.PoP/SignedHttpRequest/PopTokenValidationPolicy.cs
@@ -34,7 +34,7 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 
-namespace Microsoft.IdentityModel.Protocols.PoP
+namespace Microsoft.IdentityModel.Protocols.PoP.SignedHttpRequest
 {
     /// <summary>
     ///
